Move parameter-descriptor out of components
