+++ packages/typescript/src/components/index.ts
@@ -27,4 +27,4 @@ export * from "./TsConfigJson.js";
 export * from "./TypeDeclaration.js";
 export * from "./ValueExpression.js";
 export * from "./VarDeclaration.js";
-export * from "./ParameterDescriptor.js";
+export * from "../parameter-descriptor.js";
