+++ packages/typescript/test/function-declaration.test.tsx
@@ -2,8 +2,8 @@ import { Props, refkey } from "@alloy-js/core";
 import { d } from "@alloy-js/core/testing";
 import { describe, expect, it } from "vitest";
 import { StatementList } from "../../core/src/components/StatementList.jsx";
-import { ParameterDescriptor } from "../src/components/ParameterDescriptor.js";
 import { FunctionDeclaration, VarDeclaration } from "../src/index.js";
+import { ParameterDescriptor } from "../src/parameter-descriptor.js";
 import { toSourceText } from "./utils.js";
 
 it("works", () => {
